Move file download logic to its own helper file
